refactor: app 레이아웃 수정
라우트 및 layout 관리
