+++ src/layouts/NoLayout.js
@@ -1,0 +1,5 @@
+import { Outlet } from 'react-router-dom';
+
+export function NoLayout() {
+  return <Outlet />;
+}
